Move TaskTypes to core.tasks
