+++ packages/examples/cvat/exchange-oracle/src/core/tasks/__init__.py
@@ -1,0 +1,5 @@
+from src.core.tasks.types import TaskTypes
+
+__all__ = [
+    "TaskTypes"
+]
