+++ packages/examples/cvat/exchange-oracle/src/endpoints/exchange.py
@@ -14,7 +14,8 @@ import src.cvat.api_calls as cvat_api
 import src.services.cvat as cvat_service
 import src.services.exchange as oracle_service
 from src.core.config import Config
-from src.core.types import ProjectStatuses, TaskTypes
+from src.core.tasks import TaskTypes
+from src.core.types import ProjectStatuses
 from src.db import SessionLocal
 from src.db import engine as db_engine
 from src.endpoints.authentication import (
