+++ packages/examples/cvat/exchange-oracle/tests/integration/cron/state_trackers/test_track_completed_tasks.py
@@ -3,7 +3,8 @@ import uuid
 
 from sqlalchemy.sql import select
 
-from src.core.types import JobStatuses, Networks, ProjectStatuses, TaskStatuses, TaskTypes
+from src.core.tasks import TaskTypes
+from src.core.types import JobStatuses, Networks, ProjectStatuses, TaskStatuses
 from src.crons.cvat.state_trackers import track_completed_tasks
 from src.db import SessionLocal
 from src.models.cvat import Job, Project, Task
